refactor: Keep old ColorRampDropdown filename
